+++ trifood_javaBE/src/main/java/com/tripath/trifood/api/trifood/services/service/EGroupService.java
@@ -4,19 +4,17 @@ import com.tripath.trifood.api.trifood.dto.EGroupDto;
 import com.tripath.trifood.api.trifood.response.EGroupResponse;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 public interface EGroupService {
     EGroupDto createEGroup(EGroupDto eGroupDto);
 
-    EGroupDto updateEGroup(EGroupDto EGroupDto, Integer eGroupId);
+    EGroupDto updateEGroup(EGroupDto EGroupDto, Long eGroupId);
 
-    void deleteEGroup(Integer eGroupId);
+    void deleteEGroup(Long eGroupId);
 
     EGroupResponse getAllEGroup(Integer pageNumber, Integer pageSize, String sortBy, String sortDir);
 
-    EGroupDto getEGroupById(Integer eGroupId);
+    EGroupDto getEGroupById(Long eGroupId);
 
     EGroupResponse searchEGroup(String keyword, Integer pageNumber, Integer pageSize);
 }
